backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.3
django-debug-toolbar==4.3.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
psycopg==3.1.19
psycopg-binary==3.1.19
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1